# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LowPriceClone",
        "description": "List of endpoints",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/banniere/store": {
            "post": {
                "tags": [
                    "Banniere"
                ],
                "summary": "Create a new banniere",
                "operationId": "76e322923007a5201c483d61141a8eeb",
                "requestBody": {
                    "description": "banniere object to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "pagelink"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New banniere"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "image.jpg"
                                    },
                                    "pagelink": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "banniere created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "New banniere"
                                        },
                                        "image": {
                                            "type": "string",
                                            "format": "url",
                                            "example": "banniere/image.jpg"
                                        },
                                        "pagelink": {
                                            "type": "string",
                                            "format": "url",
                                            "example": "https://example.com/"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/banniere/update/{id}": {
            "put": {
                "tags": [
                    "Banniere"
                ],
                "summary": "Mettre à jour une Banniere",
                "operationId": "02594d14877b55170ecf867f1d328b19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du banniere",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Objet du banniere à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "pagelink"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New banniere"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "image.jpg"
                                    },
                                    "pagelink": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "vBanniere updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "New banniere"
                                        },
                                        "image": {
                                            "type": "string",
                                            "format": "url",
                                            "example": "image.jpg"
                                        },
                                        "pagelink": {
                                            "type": "string",
                                            "format": "url",
                                            "example": "https://example.com/"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Jeton d'authentification manquant ou invalide"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/banniere/delete/{id}": {
            "delete": {
                "tags": [
                    "Banniere"
                ],
                "summary": "Supprimer une banniere",
                "operationId": "0645129551ba2c382a6ed241d6cf093b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du baniiere",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banniere delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Banniere supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Jeton d'authentification manquant ou invalide"
                    },
                    "404": {
                        "description": "Banniere non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token should be provided in the 'Authorization' header. Example: Bearer {token}",
                "scheme": "bearer"
            }
        }
    }
}